requirements.txt
pyzbar==0.1.8
Pillow==7.1.2
pyotp==2.3.0
